+++ scaffolding-way/src/App.vue
@@ -5,13 +5,14 @@ import HelloWorld from './components/HelloWorld.vue'
 
 <template>
   <header>
-    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+    <img alt="Vue logo" class="logo" height="125" src="@/assets/logo.svg" width="125" />
     <div class="wrapper">
-      <HelloWorld msg="Thank you" />
+      <HelloWorld msg="Vue TodoList" />
 
       <nav>
         <RouterLink to="/">Home</RouterLink>
         <RouterLink to="/about">About</RouterLink>
+        <RouterLink to="/assignments">Assignments</RouterLink>
       </nav>
     </div>
   </header>