+++ scaffolding-way/src/mixins/flash.js
@@ -1,14 +1,14 @@
-import swal from "sweetalert";
+import swal from 'sweetalert'
 
 export default {
-    methods: {
-        flash(message, title = "Success", type = "success") {
-            return swal({
-                title: title,
-                text: message,
-                icon: type,
-                button: "OK",
-            });
-        }
+  methods: {
+    flash(message, title = 'Success', type = 'success') {
+      return swal({
+        title: title,
+        text: message,
+        icon: type,
+        button: 'OK'
+      })
     }
+  }
 }