+++ scaffolding-way/src/views/AssignmentView.vue
@@ -1,0 +1,10 @@
+<script setup>
+import Assignments from '@/components/assignments/Assignments.vue'
+</script>
+<template>
+  <main>
+    <Assignments></Assignments>
+  </main>
+</template>
+
+<style scoped></style>