+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 # Changelog
+
+## [1.0.3] - 2023-06-03
+
+### Add
+
+- Add Assignments to handle AssignmentLists
+- Create in-progress, completed AssignmentLists.
+- Add bindings between the list and it's items to update it immediately.
+
 ## [1.0.2] - 2023-06-03
-### Add 
+
+### Add
+
 - Add new composable function 'useFlash' to handle flash messages in scaffolded app.
 - Add reactivity transform options to handle the ref value.
+
 ## [1.0.1] - 2023-06-01
+
 ### Changed
+
 - Split the repo to two parts dirty-hands and scaffolded.
 
 ## [1.0.0] - 2023-05-31